resolve domains in multiple quintiles
The quintiles are by subdomain, but our system works are full domain level. To resolve, in 99% of cases I was able to find an entry for the "root" domain and delete all others. For the other 1% I manually reviewed and kept the single most well-known (subjective judgement) source and renamed it to the "root" domain.